Implemented Collins-Krandick root isolation algorithm
Based on the infallible algorithm for counting roots in a
rectangle in the complex plane, an algorithm for isolation
of complex roots of polynomials with rational coefficients
was implemented.

The algorithm computes the number of roots in a sufficiently
large initial rectangle (using Cauchy bound) and then performs
vertical and horizontal bisections and root counting until only
rectangles with exactly one root in each remain.

The algorithm isolates roots only in the upper half plane, but
excluding the real line, so real roots of the input polynomial
are discarded immediately and can be isolated using much more
efficient continued fraction approach. Conjugates are trivially
added to the resulting rectangles list by symmetry.

The problem of roots located on boundary of isolating rectangles
is resolved by counting only roots that are located on northern
and western edges, and on north-western corner. This implies
the fact that the real line is not considered for root counting
at all.

Isolating rectangles are sorted by their real component and
then by negated imaginary component, so that root and its
conjugate are located together forming a pair.

This is not the optimal implementation of Collins-Krandick
algorithms and several significant improvements are possible.