Symbol: register local assumptions
This decreases the performance of Symbol(), but allows things like this:

>>> x = Symbol('x', even=True)
>>> ask(x, Q.even)
True

To implement this, get_local_assumptions() and set_local_assumptions() hat
to be made more flexible concerning the wanted "local" scope.

Next step is to replace all the is_* stuff with ask() queries. Then it should
be possible to remove the old system completely.