Added symbolic capabilities to factor()
If the input expression to factor() is not an instance of Poly
class and generators weren't specified, then factor won't expand
the expression and will factor its components, e.g.:

In [1]: factor((x**2 + 4*x + 4)**10000000/(x**2 + 1), extension=I)
Out[1]:
       20000000
(2 + x)
───────────────
(x + ⅈ)⋅(x - ⅈ)

(previously required frac=True and hanged anyway).

This allows us to handle very large and symbolic exponents when the
input expression is a product or fraction. Cases like the following:

 factor((x + y)**100*z**2 + 2*(x + y)**100*z + (x + y)**100)

are still not supported (the above will be expanded(!) first).